Update Kotlin compiler to version 1.3.30-R41-20190410
Contains work-arounds to b/130370477

Test: ./gradlew test assemble :ui-layout:connectedAndroidTest :ui-material:connectedAndroidTest

+++ ui-r4a/src/main/java/androidx/ui/demos/CraneText.kt
@@ -567,7 +567,7 @@ fun TextDemoTextScaleFactor() {
         <CraneWrapper>
             <Text
                 text=textSpan
-                textScaleFactor=2.0.toFloat() />
+                textScaleFactor=2.0f />
         </CraneWrapper>
     </LinearLayout>
 }
